9624: Rename all components to 'component.js'
