+++ app/javascript/components/AdvancedSearchFilter/component.js
